fix: bulk update collections
